requirements.txt
selenium==3.141.0
pandas==1.1.3
bs4==0.0.1
Pillow==8.0.1
requests==2.24.0